Created a ScriptAutoCompletor interface
