+++ qupath-extension-script-editor/src/main/java/qupath/lib/gui/scripting/richtextfx/PlainAutoCompletor.java
@@ -1,0 +1,16 @@
+package qupath.lib.gui.scripting.richtextfx;
+
+import javafx.scene.input.KeyEvent;
+
+class PlainAutoCompletor implements ScriptAutoCompletor {
+
+	@Override
+	public void applyNextCompletion() {
+		// Do nothing
+	}
+
+	@Override
+	public void resetCompletion(KeyEvent e) {
+		// Do nothing
+	}
+}
